Show reference warning in normal navigation
