+++ field_friend/automations/navigation/field_navigation.py
@@ -6,13 +6,12 @@ import numpy as np
 import rosys
 from nicegui import ui
 from rosys.analysis import track
-from rosys.geometry import GeoReference, Point
-from rosys.hardware import Gnss
+from rosys.geometry import Point
 
 from ..field import Field, Row
 from ..implements.implement import Implement
 from ..implements.weeding_implement import WeedingImplement
-from .navigation import WorkflowException
+from .navigation import WorkflowException, is_reference_valid
 from .straight_line_navigation import StraightLineNavigation
 
 if TYPE_CHECKING:
@@ -324,13 +323,3 @@ class FieldNavigation(StraightLineNavigation):
                         .polar(randint(-15, 15)*0.01, self.robot_locator.pose.yaw + np.pi/2)
                     self.detector.simulated_objects.append(rosys.vision.SimulatedObject(category_name='weed',
                                                                                         position=rosys.geometry.Point3d(x=p.x, y=p.y, z=0)))
-
-
-def is_reference_valid(gnss: Gnss, *, max_distance: float = 5000.0) -> bool:
-    if GeoReference.current is None:
-        return False
-    if gnss.last_measurement is None:
-        return False
-    if gnss.last_measurement.gps_quality == 0:
-        return False
-    return gnss.last_measurement.point.distance(GeoReference.current.origin) <= max_distance
